backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
httpx==0.27.2
supabase==2.9.1
asyncpg==0.29.0
python-multipart==0.0.12
wechatpayv3==1.3.3
alipay-sdk-python>=3.7.600,<3.8.0
alibabacloud-dysmsapi20170525>=2.0.0,<3.0.0
tencentcloud-sdk-python-sms>=3.0.1100,<3.1.0
python-dotenv==1.0.1
